server/requirements.txt
fastapi
gunicorn
pymongo
pyjwt
websockets
pydantic
python-multipart
pydantic[email]